Remote Connection URL (#150)
* "Remote Connection" button and model

* share link is in working state

* works with separate account

* request works unauth

* add home router

* connection save message

* typo

* destination is immediately created

* requested changes

+++ pkg/storage/database/static/static.go
@@ -44,7 +44,7 @@ func NewStaticDatabase(conf config.Database, destinations []config.Destination, 
 		}
 
 		for _, apiKey := range destination.APIKeys {
-			err = rc.AddAPIKey(ctx, dest.ID, rc.Hash(apiKey))
+			err = rc.AddAPIKey(ctx, int64(dest.ID), rc.Hash(apiKey))
 			if err != nil {
 				return nil, err
 			}
